more changes

+++ ConicSection/src/Part4.java
@@ -3,24 +3,24 @@ import java.util.*;
 public class Part4 {
 	
 	// function to get values
-	public static ArrayList<Long> getPoints() {
-		ArrayList<Long> points = new ArrayList<Long>();
+	public static ArrayList<Double> getPoints() {
+		ArrayList<Double> points = new ArrayList<Double>();
 		Scanner sc = new Scanner(System.in); 
 		System.out.println("Enter a, b, c values:");
 		for (int i = 0; i < 3; i++) {
-			points.add(sc.nextLong());
+			points.add(sc.nextDouble());
 		}
 		System.out.println("Enter an S value:");
-		points.add(sc.nextLong());
+		points.add(sc.nextDouble());
 		sc.close();
 		System.out.println();
 		return points;
 	}
 	
 	// function that returns updated s value 
-	public static long sVal(long initialS, long digits) {
-		long s = initialS; 
-		long k = digits - 1;
+	public static double sVal(double initialS, double digits) {
+		double s = initialS; 
+		double k = digits - 1;
 		int i = 0; 
 		while (i < k) {
 			s = ((s * 10) + initialS);  
@@ -30,9 +30,9 @@ public class Part4 {
 	}
 	
 	// function that returns points with given number of digits
-	public static ArrayList<Long> updatePoints(ArrayList<Long> points, long digits, long initialS) {
-		ArrayList<Long> lst = new ArrayList<Long>(6);
-		ArrayList<Long> ret = new ArrayList<Long>(6);
+	public static ArrayList<Double> updatePoints(ArrayList<Double> points, double digits, double initialS) {
+		ArrayList<Double> lst = new ArrayList<Double>(6);
+		ArrayList<Double> ret = new ArrayList<Double>(6);
 		for (int j = 0; j < 3; j++) {
 			lst.add(points.get(j));
 			ret.add(points.get(j));
@@ -52,7 +52,7 @@ public class Part4 {
 			i++;
 			h++; 
 		}
-		long s = sVal(initialS, digits);
+		double s = sVal(initialS, digits);
 		for (int m = 0; m < 3; m++) {
 			ret.add(s - ret.get(m));
 		}
@@ -63,16 +63,16 @@ public class Part4 {
 	}
 	
 	// function that determines type of conic 
-	public static void conicType(ArrayList<Long> points, long initialS, long digits) {
+	public static void conicType(ArrayList<Double> points, double initialS, double digits) {
 		
-		long a = points.get(0); 
-		long b = points.get(1);
-		long c = points.get(2);
-		long s = sVal(initialS, digits); 
-		long p1 = ((a*a)-(a*b)+(a*c)-(a*s)-(b*b)+(b*c)+(b*s)-(c*c));
-		long p2 = ((a*a)-(a*b)-(a*c)+(b*b)+(b*c)-(b*s)-(c*c)+(c*s));
-		long p3 = ((a*a)+(a*b)-(a*c)-(a*s)+(b*b)+(b*c)-(2*b*s)+(c*c)-(c*s)+(s*s));
-		long p4 = ((a*a)+(a*b)+(a*c)-(2*a*s)-(b*b)+(b*c)+(c*c)-(2*c*s)+(s*s));
+		double a = points.get(0); 
+		double b = points.get(1);
+		double c = points.get(2);
+		double s = sVal(initialS, digits); 
+		double p1 = ((a*a)-(a*b)+(a*c)-(a*s)-(b*b)+(b*c)+(b*s)-(c*c));
+		double p2 = ((a*a)-(a*b)-(a*c)+(b*b)+(b*c)-(b*s)-(c*c)+(c*s));
+		double p3 = ((a*a)+(a*b)-(a*c)-(a*s)+(b*b)+(b*c)-(2*b*s)+(c*c)-(c*s)+(s*s));
+		double p4 = ((a*a)+(a*b)+(a*c)-(2*a*s)-(b*b)+(b*c)+(c*c)-(2*c*s)+(s*s));
 		
 		System.out.println("P1: " + p1);
 		System.out.println("P2: " + p2);
@@ -108,9 +108,9 @@ public class Part4 {
 	
 	// main driver  
 	public static void main(String args[]) {
-		ArrayList<Long> points = getPoints(); 
-		ArrayList<Long> lst = new ArrayList<Long>();
-		long s = points.remove(3);
+		ArrayList<Double> points = getPoints(); 
+		ArrayList<Double> lst = new ArrayList<Double>();
+		double s = points.remove(3);
 		for (int i = 1; i <= 10; i++) {
 			lst = updatePoints(points, i, s);
 			conicType(lst, s, i);
